export the helpers and separate the cli
